feat: introduce dependency-injected container
This change introduces a "container" component which is used for
rendering different variants of the CountrySelectSearch component
based on the state of the query.

Moreover, any behaviour that's dependent on a specific query state
such as callback handlers for keyboard navigation can simply be
passed as props to the container instead of having to juggle refs.